+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: add one node strategy (#106)
 - The Connect tasks have a `round_idx` attribute in their metadata (#101)
 - doc: add python api to documentation #105
+- fix: naming changed from FedAVG to FedAvg #114
 
 ### Improved